Move some classes around and adjust help strings
